refactor: update hero section styles
