+++ messages/en.json
@@ -34,6 +34,7 @@
 
   "Hero": {
     "title": "Find the perfect devices for you",
+    "viewMore": "View More",
     "description": "Explore the latest gadgets and upgrade to the perfect device for your needs.",
     "button": "Shopping Now"
   },
