+++ messages/ka.json
@@ -34,6 +34,7 @@
 
   "Hero": {
     "title": "საუკეთესო ტექნიკა შენთვის",
+    "viewMore": "მეტის ნახვა",
     "description": "შეიძინე შენს საჭიროებებზე მორგებული უახლესი ტექნიკა.",
     "button": "მაღაზია"
   },
